Make default porcelain presenter easier to scan, more like compact.
This is a somewhat experimental change with the goal to make it easier
to visually scan for patterns across many repositories.

It also has the advantage that the output is more similar to compact
mode.

A disadvantage is that output is a little more repetitive now. But I
hope this is outweighed by the advantages.

Resolves https://github.com/shurcooL/gostatus/issues/29#issuecomment-219237760.