Compare repo URLs for equality ignoring scheme and userinfo.
Consider the repo URL to be the same as long as the host, path and
query match. If a different schema is used (e.g., https vs ssh), then
assume it still accesses the same content.

Similarly, ignore username and password information, since the repo
content is assumed to be the same regardless of credentials used to
access.

Parse repo URL with support for SCP-like syntax (commonly used by git
for ssh urls).

Fixes #30.