Improve behavior when offline.
When offline, RemoteBranchAndRevision will always fail. Try to use
CachedRemoteDefaultBranch or NoRemoteDefaultBranch as a fallback
(better than an empty string).

Improve detection and differentiation of unreachable remote vs no
remote. When there's no remote, "!" status is still displayed. When the
remote is unreachable (because you're offline, it's blocked, or server
is down), then "?" status is displayed.