Detect unsupported VCS vs no VCS.
Go supports only 4 VCSes: git, hg, svn and bzr. It's possible for us to
know when it's one of those 4 or not, even though vcsstate only
supports git and hg.

It's better to tell the difference between a repo not being under a VCS
that Go supports at all, or if it's in a VCS that vcsstate currently
doesn't support.